Consolidate usage-extraction tests into existing test files
